+++ bestbuyfrontend/src/main/java/com/niit/bestbuyfe/controllers/SupplierController.java
@@ -2,15 +2,20 @@ package com.niit.bestbuyfe.controllers;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import com.niit.bestbuy.dao.ProductDAO;
 import com.niit.bestbuy.dao.SupplierDAO;
+import com.niit.bestbuy.model.Product;
 import com.niit.bestbuy.model.Supplier;
 
 
@@ -19,6 +24,8 @@ public class SupplierController
 {
 	@Autowired
 	SupplierDAO supplierDAO;
+	@Autowired
+	ProductDAO productDAO;
 	
 	@RequestMapping(value="/supplier")
 	public String showSupplier(Model m)
@@ -30,12 +37,20 @@ public class SupplierController
 	}
 	
 	@RequestMapping(value="/addSupplier", method=RequestMethod.POST)
-	public String addSupplier(@ModelAttribute("addSupplier") Supplier supplier, Model m)
+	public String addSupplier(@ModelAttribute("addSupplier") @Valid Supplier supplier, BindingResult result, Model m)
 	{
-		supplierDAO.add(supplier);
-		List<Supplier> supplierList=supplierDAO.listSuppliers();
-		m.addAttribute("supplierList",supplierList);
-		m.addAttribute("addSupplier",new Supplier());
+		if(result.hasErrors())
+		{
+			m.addAttribute("errors",true);
+		}
+		else
+		{
+			supplierDAO.add(supplier);
+			List<Supplier> supplierList=supplierDAO.listSuppliers();
+			m.addAttribute("supplierList",supplierList);
+			m.addAttribute("addSupplier",new Supplier());
+		}
+		
 		return "Supplier";
 	}
 	
@@ -43,7 +58,20 @@ public class SupplierController
 	public String deleteSupplier(@PathVariable("supplierId") int supplierId, Model m)
 	{
 		Supplier supplier=supplierDAO.getSupplier(supplierId);
-		supplierDAO.delete(supplier);
+		boolean productsExistWithSameSupplierId=false;
+		for(Product product : productDAO.listProducts())
+		{
+			if(product.getSupplierId()==supplier.getSupplierId())
+			{
+				productsExistWithSameSupplierId=true;
+				break;
+			}
+		}
+		if(productsExistWithSameSupplierId)
+			m.addAttribute("deleteError",true);
+		else
+			supplierDAO.delete(supplier);
+		
 		List<Supplier> supplierList=supplierDAO.listSuppliers();
 		m.addAttribute("supplierList",supplierList);
 		m.addAttribute("addSupplier",new Supplier());
